test: build (#1)
* 7.3.0

* Adding several tests for various 'pre-release' situations.

PR-URL: https://github.com/npm/node-semver/pull/316
Credit: @raphaelzulliger
Close: #316
Reviewed-by: @isaacs

* Always exclude prereleases from range maximums

In semver v4, the ranges were simplified in 2331a9e "to remove -0
everywhere". This had the unintended side effect of including the
endpoint version's prerelease versions in the resulting range, as
lt('1.0.0-0', '1.0.0') === true.

Tests are updated, adding -0 to the maximal endpoints. One test is moved
from range-include to range-exclude, comparing 2.0.0-rc1 against ^1.0.0

Fixes #223
Fixes #254

EDIT(isaacs): Updated to resolve conflicts on master branch.

PR-URL: https://github.com/npm/node-semver/pull/320
Credit: @eemeli
Close: #320
Reviewed-by: @isaacs

* 7.3.1

* Exclude prereleases in X-ranges with < comparators

This one didn't cause any bugs, but it's more for consistency and
future-proofing.

When a range like '<=1' is translated to 'anything less than 2.0.0', it
should be expressed as '<2.0.0-0' to prevent prerelease 2.0.0 versions
as well.

It didn't cause any problems with intersection or range satisfying
because it would only treat 2.0.0-pre as satisfying <2.0.0 when
includePrerelease was set, and if includePrerelease was set, then it
would add the -0 to the less-than comparator.

Nevertheless, all other areas where we use a <X.Y.0 range, we add the -0
prerelease in order to force the boundary of the set to exclude
prereleases on that version, which is good defense in depth to prevent
other errors with intersection and subsets seen previously.

* 7.3.2

* Fixed variable names and comments

PR-URL: https://github.com/npm/node-semver/pull/315
Credit: @raphaelzulliger
Close: #315
Reviewed-by: @isaacs

* Reduce unnecessary Comparators in Range set

There are a few cases that lead to Range's Comparator sets being longer
than strictly necessary.  This reduces performance of methods that
iterate over ranges repeatedly (for example, `intersects` and `subset`),
and leads to some confusing toString output like turning `x || * || X`
into `||||` instead of `*`.

- If any simple range in the set contains the null set <0.0.0-0, then
  the entire simple range is the null set. `2.x <0.0.0-0` is the same as
  just `<0.0.0-0`.  (This is used for `>*` and `<*`, which cannot match
  anything.)
- Ensure that a given Comparator will only occur once within each simple
  range set.  `2.3.x ^2.3` doesn't need to include `>=2.3.0` more than
  once.
- If a simple range set contains more than one comparator, remove any `*`
  comparators.  `* >=2.3.4` is the same as just `>=2.3.4`.  This was
  already being done in the cast to a string, but some `ANY` Comparators
  would be left behind in the set used for matching.
- If a Range set contains the simple range `*`, then drop any other
  simple ranges in the set. `* || 2.x` is the same as `*`.

There's still some unnecessary comparators in there.  For example, the
range `2.3 ^2.3.4` parses to `>=2.3.0 <2.4.0-0 >=2.3.4 <3.0.0-0`.  Of
course, anything that is `<2.4.0-0` is also `<3.0.0-0`, and anything
that is `>=2.3.4` is also `>=2.3.0`, so the `<3.0.0-0` and `>=2.3.0`
Comparators are not necessary.  But simplifying those out would be a bit
more work.

To do that, we could walk the set of Comparators checking to see if they
are a subset of any other Comparators in the list, and if so, removing
them.  The subset check would not have to be a full Range.subset(); we
could just see if the gtlt points in the same direction, and if one
semver is greater than the other.  It's an O(n^2) operation, but one on
typically very small n.

PR-URL: https://github.com/npm/node-semver/pull/324
Credit: @isaacs
Close: #324

* add a test for the coverage map

This also enforces our strict 1:1 mapping from tests to the system under test.

* Memoize range parsing

This yields about a 10-20% speed improvement to a lot of common npm
operations that spend most of their CPU time parsing a relatively small
set of dependency ranges.

It does mean that one `range.set` may contain the same Comparator
objects as another range with the same sub-range.  That was technically
already possible, due to the `Comparator.ANY` magic range, but it did
require a bit of tweaking in the `subset` function, which is also now
a bit faster when calculating subsets for ranges that are partly or
entirely identical.

* spelling: intersection

* spelling: satisfies

PR-URL: https://github.com/npm/node-semver/pull/339
Credit: @jsoref
Close: #339
Reviewed-by: @isaacs

* MinVersion failing for '^2.16.2 ^2.16' alike range

Patch for issue #340

EDIT(@isaacs): updated to make it not break other stuff, and added test.

PR-URL: https://github.com/npm/node-semver/pull/341
Credit: @alfredsgenkins
Close: #341
Reviewed-by: @isaacs

* Add lru-cache dep

Should've been added for range memoization, missed because dev-dep

* 7.3.3

* use modern lru-cache, not legacy

* 7.3.4

* fix(subset): check any as superset

Adds short-circuit check if superset is `*`.

PR-URL: https://github.com/npm/node-semver/pull/375
Credit: @jameschensmith
Close: #375
Reviewed-by: @isaacs

* Correctly handle prereleases/ANY ranges in subset

An "ANY" range (ie, `""`, `*`, etc.) does not include prerelease
versions except when `includePrerelease` flag is set.

Also, merely looking at the max/min boundaries of any ranges ignores the
fact that the sub range maybe including prerelease versions that are
excluded from the super range.  For example, `>=1.2.3-pre.0` is _not_ a
subset of `>=1.0.0`, because it inludes `1.2.3-pre.0`, `1.2.3-pre.1`,
and so on.

PR-URL: https://github.com/npm/node-semver/pull/377
Credit: @isaacs
Close: #377
Reviewed-by: @wraithgar

* 7.3.5

Co-authored-by: isaacs <i@izs.me>
Co-authored-by: Raphael Zulliger <raphael.zulliger@software-natives.ch>
Co-authored-by: Eemeli Aro <eemeli@gmail.com>
Co-authored-by: Josh Soref <jsoref@users.noreply.github.com>
Co-authored-by: Alfreds Genkins <fredy.genkin@gmail.com>
Co-authored-by: James Chen-Smith <jameschensmith@gmail.com>